Cleanup (#1935)
