+++ packages/components/src/spectrum/picker/index.ts
@@ -1,4 +1,3 @@
 export * from './Picker';
-export * from './PickerFromChildren';
-export * from './PickerFromItems';
 export * from './PickerModel';
+export * from './usePickerScrollOnOpen';
